requirements.txt
numpy==1.21.4
pandas==1.3.4
pi==0.1.2
Pillow==8.4.0
python-dateutil==2.8.2
pytz==2021.3
six==1.16.0
